Improved readability of docker-entrypoint

+++ cms-app/docker-entrypoint.sh
@@ -3,7 +3,7 @@
 echo "Starting CMS services..."
 cd /root/cms
 
-#PGPASSWORD=$POSTGRES_PASSWORD psql -d $POSTGRES_DB -h cms-db -U $POSTGRES_USER -c "SELECT EXIST (SELECT COUNT(name) FROM admins);"
+# Database & Admin account initialization
 if [ ! -f "init.lock" ]; then
    echo "DB needs initialization..."
    python3 scripts/cmsInitDB
@@ -14,22 +14,31 @@ else
    echo "DB was already initialized..."
 fi
 
-# Start Logging Service
+# Start Admin & Logging Service
 python3 scripts/cmsLogService 0   | tee -a /var/log/cms/logservice &
 python3 scripts/cmsAdminWebServer | tee -a /var/log/cms/adminserver &
-
 sleep 4
 
-# Install all course packages for all courses
+# Import all required contest data
 find /root/cms-data/ -type f -name '.itime*' -delete
 python3 cmscontrib/ImportUser.py --all -L italy_yaml /root/cms-data/
 contest_out=$(python3 cmscontrib/ImportContest.py -i -u -U -L italy_yaml /root/cms-data/)
 contest_id=($(echo $contest_out | sed -nr 's/.*\(new contest id: ([0-9]+)\)\.$/\1/p'))
-pip3 install -r /root/cms-data/requirements.txt
 
+# Install python packages which are needed for the assignments
+if [ -f "/root/cms-data/requirements.txt" ]; then
+  echo "Python package requirements for assignments found. Installing..."
+  pip3 install -r /root/cms-data/requirements.txt
+else
+  echo "No Python package requirements found for assignments."
+fi
+
+# Run Contest, Scoring & Evaluation Services to setup grading environment
 python3 scripts/cmsContestWebServer -c $contest_id  | tee -a /var/log/cms/contestserver &
 python3 scripts/cmsScoringService -c $contest_id    | tee -a /var/log/cms/scoringservice &
 python3 scripts/cmsEvaluationService -c $contest_id | tee -a /var/log/cms/evaluationservice &
+
+# Initialize worker nodes (number of workers are obtained from cms-config/cms.conf)
 worker=$(python3 scripts/getconf.py number_of_workers)
 for i in $(seq 1 $worker); do
   python3 scripts/cmsWorker $i | tee -a /var/log/cms/worker$i &
